Install desktop environment

+++ web-install.sh
@@ -2,14 +2,14 @@
 
 set -e  # Exit immediately on error
 
-# Help message
 show_help() {
     echo "Usage: $0 [options]"
     echo "Options:"
     echo "  --help       Display this help message"
     echo "  --noconfirm  Skip confirmation prompts"
+    echo "  --de         Install desktop environment"
     echo "  --[stage]    Select a specific installation stage (e.g., --05)"
-    # Add more options and descriptions as needed
+    echo "  --clean      Clean up installation when done"
 }
 
 # Check if "--help" option is provided
@@ -18,11 +18,11 @@ if [[ "$*" == *--help* || "$*" == *-h* ]]; then
     exit 0
 fi
 
-# Color variables
+# ------------------------------------------------------------------------------
+
 GREEN='\033[0;32m'
 NC='\033[0m' # No Color
 
-# Function to print colored echo statements
 print_message() {
     local color=$1
     shift
@@ -37,47 +37,20 @@ fi
 
 print_message $GREEN "Starting web-install script."
 
-print_message $GREEN "  Modifying pacman.conf..."
-
-sed -i "s/.*ParallelDownloads.*/ParallelDownloads = 5/" /etc/pacman.conf
-sed -i "s/.*Color.*/Color/" /etc/pacman.conf
-
-print_message $GREEN "  pacman.conf modified successfully."
-
 print_message $GREEN "  Downloading installer..."
-
 sudo pacman -Sy git --needed $noconfirm
 git clone https://github.com/Nachtalb/os-setup /tmp/os-setup
+wd=$(pwd)
 cd /tmp/os-setup
 
-print_message $GREEN "  Installer downloaded successfully."
+# ------------------------------------------------------------------------------
 
-print_message $RED "Running installer..."
+print_message $GREEN "Running installer..."
+./install.sh "$@"
 
-# Check if argument of the form "--[number][number]" is set
-if [[ $* == *--[0-9][0-9]* ]]; then
-  # Extract the number from the argument
-  arg=$(echo "$*" | grep -oP -- '--\K[0-9][0-9]')
-  file="install.d/${arg}_*"
-  # Check if any file matches the pattern
-  count=0
-  for f in $file; do
-    if [[ -f $f ]]; then
-      print_message $GREEN "Running $f"
-      bash "$f" "$@"
-      count=$((count + 1))
-    fi
-  done
-  # Display error if no matching file found
-  if [[ $count -eq 0 ]]; then
-    print_message $RED "No matching file found for number $arg."
-  fi
-else
-  print_message $GREEN "Running all installation files"
-  ./install.sh "$@"
+if [[ " $* " == *" --clean "* ]]; then
+    rm -rf /tmp/os-setup
 fi
 
-
-rm -rf /tmp/os-setup
-
+cd "${wd}"
 print_message $GREEN "Installation completed successfully."
